+++ CumulusMX/ThirdParty/WebUploadBlueSky.cs
@@ -12,25 +12,48 @@ using System.Threading.Tasks;
 using ServiceStack;
 using ServiceStack.Text;
 
+using Swan.Formatters;
+
 using static System.Net.Mime.MediaTypeNames;
 using static System.Net.WebRequestMethods;
 using static SQLite.SQLite3;
 
 namespace CumulusMX.ThirdParty
 {
-	internal class WebUploadBlueSky(Cumulus cumulus, string name) : WebUploadServiceBase(cumulus, name)
+	internal class WebUploadBlueSky : WebUploadServiceBase
 	{
 		private string authToken;
 		private string did;
-
-		//CancellationToken token, string userUrl = null, string userUrlPreamble = "Website: ", string baseUrl = "https://bsky.social"
-
 
 		public CancellationToken CancelToken { get; set; }
 		public string ContentTemplate { get; set; } = string.Empty;
 		public string BaseUrl { get; set; }
 		public string Language { get; set; }
+		public TimeSpan[] TimedPosts { get; set; }
+		public int TimedPostsCount { 
+			get
+			{
+				var cnt = 0;
+				for (var i = 0; i < TimedPosts.Length; i++)
+				{
+					if (TimedPosts[i] < TimeSpan.MaxValue)
+					{
+						cnt++;
+					}
+				}
+				return cnt;
+			}
+		}
 
+
+		public WebUploadBlueSky(Cumulus cumulus, string name) : base(cumulus, name)
+		{
+			TimedPosts = new TimeSpan[5];
+			for (var i = 0; i < TimedPosts.Length; i++)
+			{
+				TimedPosts[i] = TimeSpan.MaxValue;
+			}
+		}
 
 		internal override async Task DoUpdate(DateTime timestamp)
 		{
@@ -252,15 +275,15 @@ namespace CumulusMX.ThirdParty
 
 					var facet = new Facet();
 
-					int start = content.IndexOf(url, match.Index);
-
-					facet.index.byteStart = start + 1;
-					facet.index.byteEnd = start + label.Length + 1;
+					int start = match.Groups[1].Index;
 
 					// remove the URL plus pipe character from the start of the text
 					content = content.Remove(start, url.Length + 1);
 					// remove the pipe character from the end of the text
 					content = content.Remove(start + label.Length, 1);
+
+					facet.index.byteStart = GetUtf8BytePosition(content, start);
+					facet.index.byteEnd = GetUtf8BytePosition(content, start + label.Length);
 
 
 					facet.features[0].type = "app.bsky.richtext.facet#link";
@@ -268,7 +291,7 @@ namespace CumulusMX.ThirdParty
 
 					facets.Add(facet);
 
-					pos = facet.index.byteEnd;
+					pos = start + label.Length;
 
 					if (pos >= content.Length)
 					{
@@ -287,20 +310,20 @@ namespace CumulusMX.ThirdParty
 		{
 			var facets = new List<Facet>();
 
-			var regex = new Regex(@"(?:^|\s)(#[^\d\s]\S*)(?=\s|$)", RegexOptions.Compiled);
+			var regex = new Regex(@"#\w+");
 
 			foreach (Match match in regex.Matches(content)) 
 			{ 
-				if (match.Groups.Count == 2)
+				if (match.Groups.Count == 1)
 				{ 
-					string tag = match.Groups[1].Value; 
+					string tag = match.Groups[0].Value; 
 
 					var facet = new Facet();
 
-					int start = content.IndexOf(tag, match.Index);
+					int start = match.Groups[0].Index;
 
-					facet.index.byteStart = start;
-					facet.index.byteEnd = start + tag.Length + 1;
+					facet.index.byteStart = GetUtf8BytePosition(content, start);
+					facet.index.byteEnd = GetUtf8BytePosition(content, start + tag.Length);
 
 
 					facet.features[0].type = "app.bsky.richtext.facet#tag";
@@ -314,6 +337,12 @@ namespace CumulusMX.ThirdParty
 			return facets;
 		}
 
+		private static int GetUtf8BytePosition(string text, int index)
+		{
+			var substring = text[..index];
+
+			return Encoding.UTF8.GetByteCount(substring);
+		}
 
 		private sealed class AuthResponse
 		{
@@ -357,7 +386,7 @@ namespace CumulusMX.ThirdParty
 		private sealed class Record
 		{
 			public string text { get; set; }
-			public string createdAt { get; } = DateTime.UtcNow.ToString("O");
+			public string createdAt { get; } = DateTime.UtcNow.ToString("yyyy-MM-ddTHH:mm:ssZ");
 			public string[] langs { get; set; }
 
 			public Facet[] facets { get; set; }