+++ CumulusMX/WeatherStation.cs
@@ -1984,7 +1984,7 @@ namespace CumulusMX
 						_ = cumulus.APRS.DoUpdate(now);
 					}
 
-					if (cumulus.Bluesky.Enabled && (now.TimeOfDay.TotalMinutes % cumulus.Bluesky.Interval == 0) && !String.IsNullOrWhiteSpace(cumulus.Bluesky.ID) && !cumulus.Bluesky.Updating)
+					if (cumulus.Bluesky.Enabled && cumulus.Bluesky.Interval > 0 && (now.TimeOfDay.TotalMinutes % cumulus.Bluesky.Interval == 0) && !String.IsNullOrWhiteSpace(cumulus.Bluesky.ID) && !cumulus.Bluesky.Updating)
 					{
 						var parser = new TokenParser(cumulus.TokenParserOnToken)
 						{
@@ -1993,6 +1993,11 @@ namespace CumulusMX
 
 						_ = cumulus.Bluesky.DoUpdate(parser.ToStringFromString());
 					}
+
+					if (cumulus.Bluesky.Enabled && cumulus.Bluesky.TimedPostsCount > 0)
+					{
+						_ = cumulus.BlueskyTimedUpdate(now);
+					}					
 
 
 					if (cumulus.xapEnabled)